refactor personalPanel stuff
